[refactor] ItemCard에서 header 분리
