+++ client/src/Screen/02_SecondTab/Designer/waitBidListScreen.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { StyleSheet, Text, View, ScrollView, Image } from 'react-native';
-import ItemCard from '../../../Components/ListItem/itemCard';
+import ItemContent from '../../../Components/ListItem/itemContent';
 
 function WaitBidListScreen({ navigation, bidList }) {
   const [waitBidList, setWaitBidList] = useState([]);
@@ -11,7 +11,7 @@ function WaitBidListScreen({ navigation, bidList }) {
   return (
     <ScrollView style={styles.container}>
       {waitBidList.map((bid, index) => (
-        <ItemCard key={index} info={bid} navigation={navigation} screen="bid" />
+        <ItemContent key={index} info={bid} navigation={navigation} screen="bid" />
       ))}
     </ScrollView>
   );
